backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.4.2
huggingface_hub==0.19.4
transformers==4.36.2
torch==2.1.2
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.1
datasets==2.15.0
pandas==2.1.4
scipy==1.11.4
numpy==1.24.3
tqdm==4.66.1
scikit-learn==1.3.2 